애니메이션 처리
프레임당 걸리는 시간을 조절하고 싶다면 DXUTGetElapsedTime()의 반환값에 원하는 애니메이션간의 시간을 / 연산을 해주어야함

+++ TextureInfo.h
@@ -11,7 +11,7 @@ public:
 	D3DXVECTOR2 size = { 0,0 };
 	D3DXVECTOR2 center = { 0,0 };
 	D3DXVECTOR2 scale = { 1,1 };
-	int index = 0;
+	float index = 0;
 	float rotation = 0;
 	D3DXCOLOR color = D3DCOLOR_ARGB(255, 255, 255, 255);
 	bool camera_on = true;
